# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ank's Boutique API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: badge
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: inStock
          in: query
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Categories
...
  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      summary: Get wishlist by session
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Wishlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"
  /wishlist/toggle:
    post:
      operationId: toggleWishlist
      tags: [wishlist]
      summary: Toggle product in wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistToggle"
      responses:
        "200":
          description: Updated wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistToggleResponse"

  # Reviews
  /products/{id}/reviews:
    get:
      operationId: listProductReviews
      tags: [reviews]
      summary: List reviews for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a product review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # Newsletter
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, title, price, images, categoryId, inStock, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        comparePrice:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        sizes:
          type: array
          items:
            type: string
        stock:
          type: integer
        inStock:
          type: boolean
        badge:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [title, price, categoryId, images]
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        comparePrice:
          type: number
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        sizes:
          type: array
          items:
            type: string
        stock:
          type: integer
        badge:
          type: string
        sku:
          type: string
    ProductUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        comparePrice:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        categoryId:
          type: integer
        sizes:
          type: array
          items:
            type: string
        stock:
          type: integer
        badge:
          type: ["string", "null"]
        sku:
          type: string
    ProductListResponse:
      type: object
      required: [products, total, page, limit]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    WishlistItem:
      type: object
      required: [id, productId, product]
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
    WishlistToggle:
      type: object
      required: [sessionId, productId]
      properties:
        sessionId:
          type: string
        productId:
          type: integer
    WishlistToggleResponse:
      type: object
      required: [added, productId]
      properties:
        added:
          type: boolean
        productId:
          type: integer
    Review:
      type: object
      required: [id, productId, reviewerName, rating, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        reviewerName:
          type: string
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        verified:
          type: boolean
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [reviewerName, rating]
      properties:
        reviewerName:
          type: string
        reviewerEmail:
          type: string
        rating:
          type: integer
        comment:
          type: string
...