# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ank's Boutique API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  comparePrice?: number | null;
  images: string[];
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  sizes?: string[];
  stock?: number;
  inStock: boolean;
  /** @nullable */
  badge?: string | null;
  /** @nullable */
  sku?: string | null;
  /** @nullable */
  rating?: number | null;
  reviewCount?: number;
  createdAt: string;
}