# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ank's Boutique API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  title: string;
  description?: string;
  price: number;
  comparePrice?: number;
  images: string[];
  categoryId: number;
  sizes?: string[];
  stock?: number;
  badge?: string;
  sku?: string;
}