# lib/api-zod/src/generated/types/wishlistToggle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ank's Boutique API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WishlistToggle {
  sessionId: string;
  productId: number;
}